Moves various common handlers around to support adding tests for the groups events
