+++ lib/clients/rtm/event-handlers/channels.js
@@ -6,7 +6,7 @@ var zipObject = require('lodash.zipobject');
 
 var RTM_EVENTS = require('../events/rtm-events').EVENTS;
 
-var commonHandlers = require('./common');
+var baseChannelHandlers = require('./base-channel');
 var helpers = require('./helpers');
 var models = require('../../../models');
 
@@ -46,14 +46,6 @@ var handleChannelLeft = function (activeUserId, activeTeamId, dataStore, message
 };
 
 
-/** {@link https://api.slack.com/events/channel_archive|channel_archive} */
-var handleChannelArchive = helpers.setObjectProperty(true, 'isArchived');
-
-
-/** {@link https://api.slack.com/events/channel_unarchive|channel_unarchive} */
-var handleChannelUnarchive = helpers.setObjectProperty(false, 'isArchived');
-
-
 /** {@link https://api.slack.com/events/channel_rename|channel_rename} */
 var handleChannelRename = function (dataStore, message) {
     var channel = dataStore.getChannelById(message.channel.id);
@@ -62,15 +54,15 @@ var handleChannelRename = function (dataStore, message) {
 
 
 var handlers = [
-    [RTM_EVENTS.CHANNEL_ARCHIVE, handleChannelArchive],
+    [RTM_EVENTS.CHANNEL_ARCHIVE, baseChannelHandlers.handleArchive],
     [RTM_EVENTS.CHANNEL_CREATED, handleChannelCreated],
     [RTM_EVENTS.CHANNEL_DELETED, handleChannelDeleted],
     [RTM_EVENTS.CHANNEL_HISTORY_CHANGED, helpers.noopMessage],
     [RTM_EVENTS.CHANNEL_JOINED, handleChannelJoined],
     [RTM_EVENTS.CHANNEL_LEFT, handleChannelLeft],
-    [RTM_EVENTS.CHANNEL_MARKED, commonHandlers.handleChannelGroupOrDMMarked],
+    [RTM_EVENTS.CHANNEL_MARKED, baseChannelHandlers.handleChannelGroupOrDMMarked],
     [RTM_EVENTS.CHANNEL_RENAME, handleChannelRename],
-    [RTM_EVENTS.CHANNEL_UNARCHIVE, handleChannelUnarchive]
+    [RTM_EVENTS.CHANNEL_UNARCHIVE, baseChannelHandlers.handleUnarchive]
 ];
