+++ lib/clients/rtm/event-handlers/dm.js
@@ -5,7 +5,7 @@
 var zipObject = require('lodash.zipobject');
 
 var RTM_EVENTS = require('../events/rtm-events').EVENTS;
-var commonHandlers = require('./common');
+var baseChannelHandlers = require('./base-channel');
 var helpers = require('./helpers');
 var models = require('../../../models');
 
@@ -42,7 +42,7 @@ var handleDMOpen = function (dataStore, message) {
 
 var handlers = [
     [RTM_EVENTS.IM_CREATED, handleDMCreated],
-    [RTM_EVENTS.IM_MARKED, commonHandlers.handleChannelGroupOrDMMarked],
+    [RTM_EVENTS.IM_MARKED, baseChannelHandlers.handleChannelGroupOrDMMarked],
     [RTM_EVENTS.IM_OPEN, handleDMOpen],
     [RTM_EVENTS.IM_CLOSE, handleDMClose],
     [RTM_EVENTS.IM_HISTORY_CHANGED, helpers.noopMessage]
