+++ lib/clients/rtm/event-handlers/groups.js
@@ -6,21 +6,9 @@ var zipObject = require('lodash.zipobject');
 
 var RTM_EVENTS = require('../events/rtm-events').EVENTS;
 
-var commonHandlers = require('./common');
+var baseChannelHandlers = require('./base-channel');
 var helpers = require('./helpers');
 var models = require('../../../models');
-
-
-/** {@link https://api.slack.com/events/group_archive|group_archive} */
-var handleGroupArchive = function (dataStore, message) {
-    //return helpers.setContainerProperty(dataStore, message, true, 'isArchived');
-};
-
-
-/** {@link https://api.slack.com/events/group_unarchive|group_unarchive} */
-var handleGroupUnarchive = function (dataStore, message) {
-    //return helpers.setContainerProperty(dataStore, message, false, 'isArchived');
-};
 
 
 /** {@link https://api.slack.com/events/group_joined|group_joined} */
@@ -53,13 +41,13 @@ var handleGroupRename = function (dataStore, message) {
 
 
 var handlers = [
-    [RTM_EVENTS.GROUP_ARCHIVE, handleGroupArchive],
+    [RTM_EVENTS.GROUP_ARCHIVE, baseChannelHandlers.handleArchive],
     [RTM_EVENTS.GROUP_CLOSE, helpers.noopMessage],
     [RTM_EVENTS.GROUP_JOINED, handleGroupJoined],
     [RTM_EVENTS.GROUP_LEFT, handleGroupLeft],
-    [RTM_EVENTS.GROUP_MARKED, commonHandlers.handleChannelGroupOrDMMarked],
+    [RTM_EVENTS.GROUP_MARKED, baseChannelHandlers.handleChannelGroupOrDMMarked],
     [RTM_EVENTS.GROUP_OPEN, helpers.noopMessage],
-    [RTM_EVENTS.GROUP_UNARCHIVE, handleGroupUnarchive],
+    [RTM_EVENTS.GROUP_UNARCHIVE, baseChannelHandlers.handleUnarchive],
     [RTM_EVENTS.GROUP_RENAME, handleGroupRename],
     [RTM_EVENTS.GROUP_HISTORY_CHANGED, helpers.noopMessage]
 ];
