+++ lib/clients/rtm/event-handlers/helpers.js
@@ -2,21 +2,23 @@
  *
  */
 
+var models = require('../../../models');
+
 
 var noopMessage = function(dataStore, message) {
     return message;
 };
 
 
-var setObjectProperty = function(val, key) {
-    return function(dataStore, message) {
-        var obj = dataStore.getChannelGroupOrDMById(message.channel);
-        if (obj) {
-          obj[key] = val;
-        }
-    };
+/**
+ * {@link https://api.slack.com/events/team_join|team_join}
+ * {@link https://api.slack.com/events/user_change|user_change}
+ */
+var handleNewOrUpdatedUser = function (dataStore, message) {
+    var user = new models.User(message.user);
+    dataStore.setUser(user);
 };
 
 
+module.exports.handleNewOrUpdatedUser = handleNewOrUpdatedUser;
 module.exports.noopMessage = noopMessage;
-module.exports.setObjectProperty = setObjectProperty;
