+++ lib/clients/rtm/event-handlers/team.js
@@ -7,7 +7,7 @@ var zipObject = require('lodash.zipobject');
 
 var RTM_EVENTS = require('../events/rtm-events').EVENTS;
 
-var commonHandlers = require('./common');
+var helpers = require('./helpers');
 
 
 /** {@link https://api.slack.com/events/team_domain_change|team_domain_change} */
@@ -37,7 +37,7 @@ var handlers = [
     [RTM_EVENTS.TEAM_DOMAIN_CHANGE, handleTeamDomainChange],
     [RTM_EVENTS.TEAM_RENAME, handleTeamRename],
     [RTM_EVENTS.TEAM_PREF_CHANGE, handleTeamPrefChange],
-    [RTM_EVENTS.TEAM_JOIN, commonHandlers.handleNewOrUpdatedUser]
+    [RTM_EVENTS.TEAM_JOIN, helpers.handleNewOrUpdatedUser]
 ];
