+++ lib/clients/rtm/event-handlers/user.js
@@ -7,7 +7,7 @@ var zipObject = require('lodash.zipobject');
 
 var RTM_EVENTS = require('../events/rtm-events').EVENTS;
 
-var commonHandlers = require('./common');
+var helpers = require('./helpers');
 
 
 /** {@link https://api.slack.com/events/user_typing|user_typing} */
@@ -38,7 +38,7 @@ var handlePrefChange = function (activeUserId, activeTeamId, dataStore, message)
 var handlers = [
     [RTM_EVENTS.PREF_CHANGE, handlePrefChange],
     [RTM_EVENTS.USER_TYPING, handleUserTyping],
-    [RTM_EVENTS.USER_CHANGE, commonHandlers.handleNewOrUpdatedUser]
+    [RTM_EVENTS.USER_CHANGE, helpers.handleNewOrUpdatedUser]
 ];
